Support create_type_w_document
